wip: factoring command logic into new chain
